requirements.txt
google_api_python_client==2.86.0
httplib2==0.22.0
oauth2client==4.1.3
openai==0.27.7
Pillow==9.5.0
Pillow==9.5.0
protobuf==3.20.3
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
streamlit==1.22.0
